refactor: FirebaseAuth 의존성 주입을 위한 Data Source 추가
