+++ app/src/main/java/com/wasingun/seller_lounge/data/datasource/AuthFirebaseDataSource.kt
@@ -1,4 +1,4 @@
-package com.wasingun.seller_lounge.data.repository
+package com.wasingun.seller_lounge.data.datasource
 
 import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.AuthCredential
@@ -7,13 +7,13 @@ import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
 import javax.inject.Inject
 
-class AuthRepository @Inject constructor(private val firebaseAuth: FirebaseAuth) {
-
-    fun getCurrentUser(): FirebaseUser? {
+class AuthFirebaseDataSource @Inject constructor(private val firebaseAuth: FirebaseAuth) :
+    AuthDataSource {
+    override fun getCurrentUser(): FirebaseUser? {
         return firebaseAuth.currentUser
     }
 
-    fun getSignInWithCredential(authCredential: AuthCredential): Task<AuthResult> {
+    override fun getSignInWithCredential(authCredential: AuthCredential): Task<AuthResult> {
         return firebaseAuth.signInWithCredential(authCredential)
     }
 }
