+++ app/src/main/java/com/wasingun/seller_lounge/di/AuthDataSourceModule.kt
@@ -1,0 +1,19 @@
+package com.wasingun.seller_lounge.di
+
+import com.google.firebase.auth.FirebaseAuth
+import com.wasingun.seller_lounge.data.datasource.AuthDataSource
+import com.wasingun.seller_lounge.data.datasource.AuthFirebaseDataSource
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AuthDataSourceModule {
+
+    @Provides
+    fun provideFirebaseAuth(firebaseAuth: FirebaseAuth): AuthDataSource {
+        return AuthFirebaseDataSource(firebaseAuth)
+    }
+}
